only extract simple summary when the fip entry is expanded
